feat: complete rearrange ValueResolver to interface (component 1)
Extracted ValueResolver as an interface with public methods get() and getPath().
Moved the original implementation logic to RecursiveValueResolver which implements
the new interface. Updated all production code (DistributionScope, SingleThreadEventHandlerMachine,
JoshValueVisitor, JoshTypesUnitsVisitor) and test code to use RecursiveValueResolver at
construction sites while preferring the ValueResolver interface type elsewhere. Fixed the
isQueryForPatch() string matching in SingleThreadEventHandlerMachine to recognize
RecursiveValueResolver in the toString() output. All tests and style checks pass.
